Move pkg/template/registry to pkg/template/registry/template
