+++ pkg/template/registry/template/etcd/etcd_test.go
