+++ pkg/template/registry/template/strategy.go
@@ -1,4 +1,4 @@
-package registry
+package template
 
 import (
 	"fmt"
